+++ src/PileUpPolicy.php
@@ -9,14 +9,10 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Stash;
 
-interface PileUpPolicy
+enum PileUpPolicy: string
 {
-    public const IGNORE = 'ignore';
-    public const PREEMPT = 'preempt';
-    public const SLEEP = 'sleep';
-    public const VALUE = 'value';
-
-    public const KEYS = [
-        'ignore', 'preempt', 'sleep', 'value'
-    ];
+    case IGNORE = 'ignore';
+    case PREEMPT = 'preempt';
+    case SLEEP = 'sleep';
+    case VALUE = 'value';
 }